finalised rename and dropped unused context
