+++ meal-ui/src/pages/Layout.jsx
@@ -1,0 +1,14 @@
+import MinimalistSidebar from 'components/MinimalistSidebar';
+import Footer from 'components/Navigation';
+
+const Layout = ({ children }) => {
+	return (
+		<div>
+			<MinimalistSidebar />
+			<main>{children}</main>
+			{/* <Footer className='footer-grid' /> */}
+		</div>
+	);
+};
+
+export default Layout;
